webserver/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
redis==5.0.8
pillow==10.4.0
numpy==1.26.4
python-multipart==0.0.6